Attempt to implement block-less -map (tests failing)
The main thing causing a problem here is that we don't have implicit
returns in ObjC, so if I have a -map enumerator expecting a block that
returns a value, and then call -each: on it and pass a block that
doesn't return anything, I'm breaking the contract.

Need to investigate a way to dynamically type the blocks and have
optional return statements.