Dispose of EKFibers when an enumeration is deallocated
Fixes #1. When `-find` returns, its `EKEnumerator` is deallocated,
which will cancel the in-progress enumeration on the backing `EKFiber`.

To do: Attempt to dispatch enumeration operations to the same queue so
that a separate thread is not spawned for every enumeration. Currently
there is a soft limit on the number of alive `EKFiber` instances. This
mainly manifests when spawning a large number of enumerations in a
tight loop.