EKSemaphore: improve reliability and semantics of counting semaphore test
See https://github.com/allending/Kiwi/issues/321. Adding verifiers from
multiple threads is not supported in Kiwi.

Also changed the semantics of the test case to be about waiting on a
non-boolean counting semaphore (i.e., with a value of 2, the third
thread to wait will block until the semaphore is signalled).